Remove AppConfiguration
