--- agent/agent-core/src/main/java/com/sbss/bithon/agent/core/config/AppConfiguration.java
@@ -1,48 +1,0 @@
-/*
- *    Copyright 2020 bithon.cn
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package com.sbss.bithon.agent.core.config;
-
-import com.sbss.bithon.agent.core.config.validation.NotBlank;
-
-/**
- * @author frankchen
- */
-@ConfigurationProperties(prefix = "application")
-public class AppConfiguration {
-
-    @NotBlank(message = "[%s] is blank")
-    private String env;
-
-    @NotBlank(message = "[%s] is blank")
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEnv() {
-        return env;
-    }
-
-    public void setEnv(String env) {
-        this.env = env;
-    }
-}
