+++ agent/agent-core/src/main/java/com/sbss/bithon/agent/core/starter/AgentStarter.java
@@ -18,7 +18,6 @@ package com.sbss.bithon.agent.core.starter;
 
 import com.sbss.bithon.agent.bootstrap.loader.AgentClassLoader;
 import com.sbss.bithon.agent.core.aop.InstrumentationHelper;
-import com.sbss.bithon.agent.core.config.AppConfiguration;
 import com.sbss.bithon.agent.core.context.AgentContext;
 import com.sbss.bithon.agent.core.plugin.PluginInterceptorInstaller;
 import shaded.org.apache.log4j.xml.DOMConfigurator;
@@ -89,10 +88,8 @@ public class AgentStarter {
     }
 
     private void ensureApplicationTempDirectory(AgentContext context) {
-        AppConfiguration appConfiguration = context.getAgentConfiguration().getConfig(AppConfiguration.class);
-
         File tmpDir = new File(context.getAgentDirectory() + separator + AgentContext.TMP_DIR + separator +
-                               appConfiguration.getName());
+                               context.getAppInstance().getAppName());
 
         if (!tmpDir.exists()) {
             tmpDir.mkdirs();
